requirements.txt
Django~=4.2.19
pgvector~=0.4.0
dotenv~=0.9.9
openai~=1.63.2
python-dotenv~=1.1.0
numpy~=1.26.4
djangorestframework~=3.15.2